chore: rename sm100-specific cutedsl kernels
