+++ benchmarks/bench_linear_attn.py
@@ -26,7 +26,7 @@ from fla.ops.linear_attn.utils import normalize_output
 # from fla.ops.linear_attn.naive import naive_recurrent_linear_attn
 from fla.utils import assert_close, device
 
-from cula.ops.linear_attn import LinearAttentionChunkwise
+from cula.ops.linear_attn_sm100 import LinearAttentionChunkwise
 
 os.environ.setdefault("FLA_USE_FAST_OPS", os.getenv("CULA_USE_FAST_MATH", "1"))  # Enable fast ops in FLA for fair comparison
