+++ cula/kda/__init__.py
@@ -13,11 +13,13 @@
 # limitations under the License.
 
 from cula.kda.chunk import chunk_kda
+from cula.kda.fused_fwd_sm100_wip import flash_kda_prefill as flash_kda_prefill_sm100_wip
 from cula.kda.hopper_fused_fwd import cula_kda_prefill as kda_prefill_hopper
 from cula.ops.kda_decode import fused_sigmoid_gating_delta_rule_update, kda_decode
 
 __all__ = [
     "chunk_kda",
+    "flash_kda_prefill_sm100_wip",
     "kda_decode",
     "fused_sigmoid_gating_delta_rule_update",
     "kda_prefill_hopper",
