+++ pyproject.toml
@@ -80,8 +80,8 @@ force-sort-within-sections = false
 "__init__.py" = ["F401"]
 "cula/_version.py" = ["UP007"]
 # TODO: fix undefined names (exp_g, chunk_kda_bwd_dqkwg) — WIP code
-"cula/ops/kda_fully_fused_wip.py" = ["F821"]
-"cula/kda/blackwell_fused_fwd.py" = ["F821"]
+"cula/ops/kda_fully_fused_sm100_wip.py" = ["F821"]
+"cula/kda/fused_fwd_sm100_wip.py" = ["F821"]
 
 [tool.setuptools_scm]
 # write generated version into package for runtime access
